Isolate all Cloudwok js
